backend/requirements.txt
opencv-python 
pillow 
pytesseract 
numpy 
fastapi 
uvicorn
streamlit
easyocr
python-multipart
spacy
langchain 
openai 
python-dotenv
streamlit 
pdf2image
langchain-together 
langchain-community
torchvision